fastapi_ai_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv>=1.0.0
mcp-use
mcp>=1.8.1
openai>=1.0.0
asyncpg>=0.29.0
psycopg2-binary
python-slugify
bcrypt>=3.2.0,<4.0.0
passlib[bcrypt]
python-jose[cryptography]
eth-account
resend
